% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtab.R
\name{newtab}
\alias{newtab}
\title{Creates a new Rmarkdown tab.}
\usage{
newtab(title = "Tab", header = 2, nonumber = T, subtabs = F, pills = F)
}
\arguments{
\item{title}{Tab title}

\item{header}{An int indicating Rmd html header level (e.g. 3 means "###")}

\item{nonumber}{A boolean for whether tab should be numbered or not}

\item{subtabs}{A boolean for whether tab has subtabs below it or not}

\item{pills}{A boolean for whether tab should be pill-shaped or not}
}
\description{
Make sure that Rmd chunk option results='asis' is present.
}
